requirements.txt
websockets>=12.0
httpx>=0.27
orjson>=3.9
pyyaml>=6.0
